refactored webviews slightly

+++ src/ci/discover.rs
@@ -2,7 +2,7 @@ use crate::ci::test::Outcome;
 
 #[derive(Clone, Debug)]
 pub struct Row {
-	pub number: i64,
+	pub number: usize,
 	pub solution: Outcome,
 	pub fitness: i64,
 	pub input: String,
